FIX: Distributed compilation

+++ src/config.hpp
@@ -1,7 +1,7 @@
 namespace config {
 namespace geometry {
 /** accuracy of the geometry representation, choose float or double */
-using precision_t = float;
+using precision_t = double;
 } // namespace geometry
 } // namespace config
